Better volume mount error test
